refs #79977 Add Rename button logic for search page - Move display modal from list item to list component.
