+++ packages/app/src/components/SearchPage/SearchResultList.tsx
@@ -1,7 +1,8 @@
-import React, { FC } from 'react';
+import React, { FC, useState } from 'react';
 import SearchResultListItem from './SearchResultListItem';
 import { IPageHasId } from '../../interfaces/page';
 import PaginationWrapper from '../PaginationWrapper';
+import PageRenameModal from '../PageRenameModal';
 
 // TOOD: retrieve bookmark count and add it to the following type
 export type ISearchedPage = IPageHasId & {
@@ -26,6 +27,49 @@ type Props = {
 const SearchResultList: FC<Props> = (props:Props) => {
   const { focusedPage } = props;
   const focusedPageId = focusedPage != null && focusedPage._id != null ? focusedPage._id : '';
+
+  const [isPageRenameModalShown, setIsPageRenameModalShown] = useState(false);
+  const [controlTargetPage, setControlTargetPage] = useState(focusedPage || {
+    _id: null,
+    path: null,
+    revision: {
+      _id: null,
+    },
+  });
+
+  function openPageRenameModalHandler() {
+    setIsPageRenameModalShown(true);
+  }
+
+  function renderModals() {
+
+    return (
+      <>
+        <PageRenameModal
+          isOpen={isPageRenameModalShown}
+          onClose={() => { setIsPageRenameModalShown(false) }}
+          pageId={controlTargetPage._id}
+          revisionId={controlTargetPage.revision._id}
+          path={controlTargetPage.path}
+        />
+        {/* <PageDuplicateModal
+          isOpen={isPageDuplicateModalShown}
+          onClose={closePageDuplicateModalHandler}
+          pageId={pageId}
+          path={path}
+        /> */}
+        {/* <PageDeleteModal
+          isOpen={isPageDeleteModalShown}
+          onClose={closePageDeleteModalHandler}
+          pageId={pageId}
+          revisionId={revisionId}
+          path={path}
+          isAbleToDeleteCompletely={isAbleToDeleteCompletely}
+        /> */}
+      </>
+    );
+  }
+
   return (
     <>
       {props.pages.map((page) => {
@@ -34,6 +78,8 @@ const SearchResultList: FC<Props> = (props:Props) => {
             key={page._id}
             page={page}
             onClickInvoked={props.onClickInvoked}
+            onClickControlDropdown={setControlTargetPage}
+            onClickPageRenameBtnInvoked={openPageRenameModalHandler}
             isSelected={page._id === focusedPageId || false}
           />
         );
@@ -48,7 +94,7 @@ const SearchResultList: FC<Props> = (props:Props) => {
           />
         </div>
       )}
-
+      { renderModals() }
     </>
   );
